Refactor map components.
